Missing i18n and zh-CN translation

+++ extra/update-language-files/index.js
@@ -26,10 +26,12 @@ const copyRecursiveSync = function (src, dest) {
     }
 };
 
-console.log("Arguments:", process.argv)
+console.log("Arguments:", process.argv);
 const baseLangCode = process.argv[2] || "en";
 console.log("Base Lang: " + baseLangCode);
-fs.rmdirSync("./languages", { recursive: true });
+if (fs.existsSync("./languages")) {
+    fs.rmdirSync("./languages", { recursive: true });
+}
 copyRecursiveSync("../../src/languages", "./languages");
 
 const en = (await import("./languages/en.js")).default;
@@ -39,7 +41,7 @@ console.log("Files:", files);
 
 for (const file of files) {
     if (!file.endsWith(".js")) {
-        console.log("Skipping " + file)
+        console.log("Skipping " + file);
         continue;
     }
